+++ test/type_construction_test.rb
@@ -4368,4 +4368,40 @@ EOF
     assert_equal Types::Name.new_interface(name: :_Boolean), type_env.lvar_types[:a]
     assert_equal Types::Name.new_interface(name: :_Boolean), type_env.lvar_types[:b]
   end
+
+  def test_empty_body_method
+    source = parse_ruby(<<EOF)
+class EmptyBodyMethod
+  def foo
+  end
+end
+EOF
+
+    typing = Typing.new
+    annotations = source.annotations(block: source.node)
+    checker = new_subtyping_checker(<<-EOF)
+class EmptyBodyMethod
+  def foo: () -> String
+end
+    EOF
+    const_env = ConstantEnv.new(builder: checker.builder, current_namespace: nil)
+    type_env = TypeEnv.build(annotations: annotations,
+                             subtyping: checker,
+                             const_env: const_env,
+                             signatures: checker.builder.signatures)
+
+    construction = TypeConstruction.new(checker: checker,
+                                        source: source,
+                                        annotations: annotations,
+                                        type_env: type_env,
+                                        block_context: nil,
+                                        self_type: Types::Name.new_instance(name: "::Object"),
+                                        method_context: nil,
+                                        typing: typing,
+                                        module_context: nil,
+                                        break_context: nil)
+    construction.synthesize(source.node)
+
+    refute_empty typing.errors
+  end
 end